mod6_labs/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
flet==0.28.3
flet-desktop==0.28.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
oauthlib==3.3.1
propcache==0.4.1
python-dotenv==1.2.1
repath==0.9.0
six==1.17.0
sniffio==1.3.1
yarl==1.22.0
